feat: split the best checkpointer example into two files
